requirements.txt
django
django-modeltranslation
djangocms-text-ckeditor==2.4.3
bleach
